refactor(edge): split raw state reconciliation
